Refine naming, make tests happy
