+++ Sources/HarmonicNetworkCore/ChordCluster.swift
@@ -9,16 +9,4 @@ import DataStructures
 import Pitch
 import SpelledPitch
 
-typealias ChordClusterNode = Tree<String,String>
-
-extension ChordClusterNode {
-    func printAll() {
-        switch self {
-        case let .leaf(value):
-            print(value)
-        case let .branch(value, branches):
-            print("\(value):")
-            branches.forEach { $0.printAll() }
-        }
-    }
-}
+public typealias ChordClusterNode = Tree<String,String>
