+++ Sources/HarmonicNetworkCore/ChordClusterView.swift
@@ -14,20 +14,20 @@ import Pitch
 import SpelledPitch
 import Geometry
 
-struct ChordClusterView {
+public struct ChordClusterView {
     var position: Point
     var label: String
 }
 
-struct ChordView {
+public struct ChordView {
     var position: Point
     var chord: String
 }
 
-typealias ChordClusterViewNode = Tree<ChordClusterView,ChordView>
+public typealias ChordClusterViewNode = Tree<ChordClusterView,ChordView>
 
 extension ChordClusterNode {
-    func layout(at position: Point = .zero, angle: Angle = .zero, spread: Double)
+    public func layout(at position: Point = .zero, angle: Angle = .zero, spread: Double)
         -> ChordClusterViewNode
     {
         func traverseToLayout(
@@ -62,7 +62,7 @@ extension ChordClusterNode {
 }
 
 /// - Returns: An array of points spread equidistantly around the given `centroid`.
-func pointsAndAngles(
+public func pointsAndAngles(
     count: Int,
     distance: Double,
     from centroid: Point,
